- Fixed a number of warnings. - Changed the ANTLR grammar parser to use more unique token names as some of them conflicted with other names. - TUT: updated tests for cache and parser to work with the changed implemenation. - TUT: removed existing auto completion tests, as these must be reimplemented differently. - Win: updated PrepareOutputDir.cmd to copy the needed MySQL.g and .tokens files.
[TESTS]
Ran auto completion cache + parser tests.

+++ library/mysql.parser/mysql-parser-common.cpp
@@ -199,7 +199,7 @@ uint32_t MySQLRecognitionBase::get_keyword_token(const std::string &keyword)
 {
   if (keywords.size() == 0)
   {
-    for (size_t i = 4; i <= ZEROFILL_SYMBOL; ++i)
+    for (uint32_t i = 4; i <= ZEROFILL_SYMBOL; ++i)
     {
       std::string name((char*)MySQLParserTokenNames[i]);
       if (base::ends_with(name, "_SYMBOL"))
